% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMOUNTAIN.R
\name{multilayernetworkSimulation}
\alias{multilayernetworkSimulation}
\title{Illustration of multi-layer weighted network simulation}
\usage{
multilayernetworkSimulation(n, k, theta, L)
}
\arguments{
\item{n}{number of nodes in each layer of the network}

\item{k}{number of nodes in the conserved module}

\item{theta}{module node score follow the uniform distribution in range [theta,1]}

\item{L}{number of layers}
}
\value{
a list containing all the layers, each as result object of \link{networkSimulation}
}
\description{
Simulate a multi-layer weighted network with each layer sharing the same set
of nodes but different nodes scores
}
\examples{
n = 100
k = 20
theta = 0.5
L = 5
cpl <- multilayernetworkSimulation(n,k,theta,L)
## No proper way to visualize it yet
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\seealso{
\code{\link{networkSimulation}}
}
\keyword{simulation}

