% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMOUNTAIN.R
\name{vecconsensus}
\alias{vecconsensus}
\title{Vector consensus}
\usage{
vecconsensus(Mat)
}
\arguments{
\item{Mat}{membership matrix from other layers, n x l matrix}
}
\value{
n-length vector
}
\description{
a consensus of several vectors, keeping nonzero positions in all vectors
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}

