% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMOUNTAIN.R
\name{moduleIdentificationGPFixSSMultilayer}
\alias{moduleIdentificationGPFixSSMultilayer}
\title{Module Identification for multi-layer network}
\usage{
moduleIdentificationGPFixSSMultilayer(W, zi, xrest, x0, lambdai = 1,
  lambda = 1, maxiter = 1000, a = 0.5)
}
\arguments{
\item{W}{edge score matrix of the network, n x n matrix}

\item{zi}{node score vector of this layer, n-length vector}

\item{xrest}{consensus solution from the rest membership \code{\link{vecconsensus}}}

\item{lambda}{parameter in objective, coefficient of node score of network}

\item{maxiter}{maximal interation of whole procedure}

\item{a}{parameter in elastic net the same as in \code{\link{EuclideanProjectionENNORM}}}
}
\value{
a list containing solution for network 1 and network 2
}
\description{
Algorithm for Module Identification on multi-layer network sharing the same set of genes
}
\examples{
n=100
k=20
theta = 0.5
pp <- networkSimulation(n,k,theta)
W <- pp[[1]]
moduleid <- pp[[3]]
z1 <- runif(n, min=0, max=1)
z1[moduleid] <- runif(k, min=theta, max=1)
z2 <- runif(n, min=0, max=1)
z2[moduleid] <- runif(k, min=theta, max=1)
z3 <- runif(n, min=0, max=1)
z3[moduleid] <- runif(k, min=theta, max=1)
x0=rep(1/n,n)
x1=x0
x2=x0
x3=x0
xrest = vecconsensus(cbind(x2,x3))
x1 = moduleIdentificationGPFixSSMultilayer(W,z1,xrest,x0,lambdai=1,lambda=1,maxiter=100,a=0.5)
xrest = vecconsensus(cbind(x1,x3))
x2 = moduleIdentificationGPFixSSMultilayer(W,z2,xrest,x0,lambdai=1,lambda=1,maxiter=100,a=0.5)
xrest = vecconsensus(cbind(x1,x2))
x3 = moduleIdentificationGPFixSSMultilayer(W,z3,xrest,x0,lambdai=1,lambda=1,maxiter=100,a=0.5)
}
\author{
Dong Li, \email{dxl466@cs.bham.ac.uk}
}
\references{
MOUNTAIN
}
\seealso{
\code{\link{vecconsensus}}
}
\keyword{identification,}
\keyword{module}
\keyword{two-layer}

